myers1999: Improve the caching mechanism of bit patterns.
I think I have found a way to avoid the cache-collision problem.

Look here:

    Total : 10000 calls
    Inputs: [\u0000...\u1fff]

    #      n=64     n=128    n=256    n=512    n=1024
    Before 0.072    0.310    1.225    4.842    19.397
    After  0.017    0.070    0.245    0.906    3.538

The break-through was the realization that we only needed to hold 64
patterns at most to compute a row of the DP matrix. Why? Because the
shorter string is guaranteed to be less than 64-characters long and,
for these characters that are not contained in the string, their bit
patterns are always 0^64.

Note that this patch implements a simple open-addressing hash table on
Peq[256] to hold these precomputed patterns. This should be as much as
efficient as the previous design, even for short strings.