Optimize mbleven for major use cases
This is my attempt to improve the performance of mbleven for the
most common use cases.

In particular, I added a special function for ASCII strings, since
we handle ASCII strings a lot. I also encoded the model table into
the (more efficient) binary representation.

Here is a benchmark result on Intel Xeon E5-2660.

BEFORE:

    polyleven.levenshtein (k=3)    0.276           3599047
    polyleven.levenshtein (k=2)    0.231           4293309
    polyleven.levenshtein (k=1)    0.217           4563204

AFTER

    polyleven.levenshtein (k=3)    0.243           4073851
    polyleven.levenshtein (k=2)    0.225           4416767
    polyleven.levenshtein (k=1)    0.215           4619078

Signed-off-by: Fujimoto Seiji <fujimoto@ceptord.net>